+++ frontend/src/components/atoms/SignOut.tsx
@@ -1,40 +1,10 @@
-import { FC, useEffect, useState } from 'react';
-import { Link, useHistory } from 'react-router-dom';
+import { FC } from 'react';
+import { Link } from 'react-router-dom';
 import { Header, Grid, Segment } from 'semantic-ui-react';
-import { Fetchsessiondestroy } from 'apis/Session';
-import { useResetRecoilState } from 'recoil';
-import LoginState from 'atom';
-
-type State = {
-  logout: boolean;
-};
+import useSignout from 'hooks/Signout';
 
 const SignOut: FC = () => {
-  const [state, setState] = useState<State>({ logout: false });
-  const history = useHistory();
-  const resetUser = useResetRecoilState(LoginState);
-
-  useEffect(() => {
-    let isMounted = true;
-    if (isMounted) {
-      Fetchsessiondestroy()
-        .then((result) =>
-          result !== undefined && result === 200
-            ? (resetUser(),
-              setState((prevUser) => ({ ...prevUser, logout: true })))
-            : history.push('/', {
-                message: 'ログアウトに失敗しました。',
-                type: 'error',
-              }),
-        )
-        .catch(() => setState({ logout: false }));
-      setTimeout(() => history.push('/'), 6000);
-    }
-
-    return (): void => {
-      isMounted = false;
-    };
-  }, [history, resetUser]);
+  const state = useSignout();
 
   return (
     <>
